% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{bin.prioritize}
\alias{bin.prioritize}
\title{Binned sorting method}
\usage{
bin.prioritize(peak.df, binwidth = 0.01)
}
\arguments{
\item{peak.df}{Peak data frame to sort}

\item{binwidth}{Width of the bins from which to pick peaks}
}
\value{
Data frame of peaks
}
\description{
Reorders peaks using the binned method.
}
\details{
Binned method works as follows:
\enumerate{
\item Retention time is divided into bins of specified binwidth.
\item Most intense peaks from each bin are arranged in order of decreasing intensity.
\item Second most intense peaks from each bin are arranged in order of decreasing intensity.
\item Process repeated until all peaks are sorted.
}
Note: in this implementation, the rownames of original peak data frame will be removed in output.
}
\seealso{
\code{\link{intensity.prioritize}}, \code{\link{random.prioritize}}
}
