% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binned_method_cpp_wrap}
\alias{binned_method_cpp_wrap}
\title{Internal function called as part of the binned algorithm. Not meant for the
end user.}
\usage{
binned_method_cpp_wrap(bin_id, intensity)
}
\arguments{
\item{bin_id}{A vector where bin_id[i] represents the bin number of peak i.}

\item{intensity}{A vector where intensity[i] represents the intensity of peak i.}
}
\value{
Vector of i indicating the order of peaks as sorted by the binned method
}
\description{
In the following, i is 0, 1, 2, ...
}
