% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{boxplot2}
\alias{boxplot2}
\title{Wrapper for boxplot function that also prints P values}
\usage{
boxplot2(y.list, ...)
}
\arguments{
\item{y.list}{a named list of vectors indicating values for each group}

\item{...}{additional arguments to the boxplot function}
}
\value{
a boxplot with P values
}
\description{
Wrapper for boxplot function that also prints P values
}
\examples{
    boxplot2(list(`Foo`=1:3, `Bar`=10:13))
}
