% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{check.peak.df}
\alias{check.peak.df}
\title{Validate peak data frame}
\usage{
check.peak.df(peak.df)
}
\arguments{
\item{peak.df}{Peak data frame}
}
\value{
TRUE if OK otherwise FALSE
}
\description{
Checks peak data frame to make sure it's valid
}
