% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{cluster.prioritize}
\alias{cluster.prioritize}
\title{Clustering peak sorting method}
\usage{
cluster.prioritize(peak.df, rt.diff = 0.05)
}
\arguments{
\item{peak.df}{Peak data frame}

\item{rt.diff}{Retention time difference between peak pairs in minutes
(default 0.05 min.)}
}
\value{
Data frame of sorted peaks
}
\description{
Ranks the peaks using a clustering algorithm. Experimental.
}
\details{
Algorithm works as follows:
\enumerate{
\item Find pairs of peaks that have retention time difference
   below a given threshold (default 0.05 min.)
\item Find clusters from the edge list constructed between peak pairs
\item Peaks are ranked by decreasing intensity within each cluster
\item If peaks have equal rank in their respective clusters, then
   re-ranked from highest to lowest intensity
}
}
