% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{dwell.time.boxplot}
\alias{dwell.time.boxplot}
\title{Boxplots of dwell times of multiple peak lists}
\usage{
dwell.time.boxplot(peak.df.list, xlab = "Method",
  ylab = "Dwell time (min.)", ...)
}
\arguments{
\item{peak.df.list}{a list of peak data frames}

\item{xlab}{label for X axis}

\item{ylab}{label for y axis}

\item{...}{additional parameters for boxplot}
}
\value{
boxplot of dwell times where each box corresponds to each peak data frame in list
}
\description{
Plots a boxplot of dwell times of different peak lists
}
