% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{dwell.times}
\alias{dwell.times}
\title{Dwell times of peaks in data frame}
\usage{
dwell.times(peak.df)
}
\arguments{
\item{peak.df}{Peak data frame}
}
\value{
A vector of retention time differences
}
\description{
Finds the dwell time of a data frame containing peak data
(defined here as the retention time difference
between successive peaks in the data frame).
}
