% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\docType{data}
\name{idh1}
\alias{idh1}
\title{Idh1 knock-in mouse dataset}
\format{A data frame with 5000 rows and 4 variables:
\describe{
\item{Rt}{Retention time in minutes}
\item{Mz}{Mass to charge ratio}
\item{Intensity}{Normalized intensity}
\item{Name}{Unique peak ID}
}}
\usage{
idh1
}
\description{
Peak data frame containing peaks from untargeted LC-MS profiling
of Idh1 knock-in mouse plasma on C18 column negative mode.
For more details, please see (unpublished paper/thesis)
}
\seealso{
\code{\link{idh1.true.pos}}
}
\keyword{datasets}
