% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\docType{data}
\name{idh1.true.pos}
\alias{idh1.true.pos}
\title{True positive peaks in the Idh1 knock-in metabolomics dataset}
\format{An object of class \code{character} of length 644.}
\usage{
idh1.true.pos
}
\description{
A list of peak IDs from the Idh1 knock-in metabolomics experiment
whose m/z correspond to metabolites in the HMDB database.
For more details, please see (unpublished paper/thesis)
}
\seealso{
\code{\link{idh1}}
}
\keyword{datasets}
