% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{intensity.prioritize}
\alias{intensity.prioritize}
\title{Sorts peak data frame by intensity}
\usage{
intensity.prioritize(peak.df)
}
\arguments{
\item{peak.df}{Peak data frame}
}
\value{
The peak data frame arranged in descending order of intensity
}
\description{
Given a peak data frame, reorders data frame
by peak intensity (maximum to minimum intensity). Note: rownames will be removed in output.
}
\seealso{
\code{\link{bin.prioritize}}, \code{\link{random.prioritize}}
}
