% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{peak.plot.df}
\alias{peak.plot.df}
\title{Generates graphable data from a list of peak data frames}
\usage{
peak.plot.df(peak.df.list)
}
\arguments{
\item{peak.df.list}{a list of peak data frames. Names indicate the method to be labelled on plot.}
}
\value{
a data frame for graphing
}
\description{
Note: intended to be called by other graphing functions
}
