% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{peak.plot}
\alias{peak.plot}
\title{Creates plots of peaks across multiple sorted peak data frames}
\usage{
peak.plot(peak.df.list)
}
\arguments{
\item{peak.df.list}{a list of peak data frames}
}
\value{
plots showing intensity vs. retention time of each peak
}
\description{
Generates plots of the peaks in peak lists
}
