% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{random.prioritize}
\alias{random.prioritize}
\title{Randomly change order of peaks in data frame}
\usage{
random.prioritize(peak.df)
}
\arguments{
\item{peak.df}{Peak data frame}
}
\value{
Data frame with rows in random order
}
\description{
Returns peaks in random order. Note: rownames will be removed in output.
}
\seealso{
\code{\link{bin.prioritize}}, \code{\link{intensity.prioritize}}
}
