% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{rank.fun}
\alias{rank.fun}
\title{Finds rank of values in descending order}
\usage{
rank.fun(x)
}
\arguments{
\item{x}{a vector to be sorted}
}
\value{
sorted vector
}
\description{
Ranking function called by various prioritization methods
}
\details{
Orders values in descending order
}
