% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{read.peak.file}
\alias{read.peak.file}
\title{Read peak data from file}
\usage{
read.peak.file(filename)
}
\arguments{
\item{filename}{The file containing peak data
(exactly four tab-separated columns:
retention time (min.), m/z, intensity, peak name)
No header row.}
}
\value{
data frame of peaks
}
\description{
Reads peak data from given text file
}
