% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{roc.plot}
\alias{roc.plot}
\title{Receiver operating characteristic curves of multiple sorted peak data frames}
\usage{
roc.plot(peak.df.list, true.pos.peaks)
}
\arguments{
\item{peak.df.list}{a list of peak data frames}

\item{true.pos.peaks}{a vector containing names of true positive peaks}
}
\value{
plot of multiple ROC curves
}
\description{
Make multiple receiver operating characteristic curves
}
\seealso{
\code{\link{roc.auc}}
}
