% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{validate.peak.df.list}
\alias{validate.peak.df.list}
\title{Validate a list of peak data frames}
\usage{
validate.peak.df.list(peak.df.list)
}
\arguments{
\item{peak.df.list}{list of peak data frames to test}
}
\value{
TRUE if all data frames are valid. Otherwise, FALSE
}
\description{
Checks whether all objects in list are valid peak data frames
}
\seealso{
\code{\link{check.peak.df}}
}
